+++ CHANGELOG.md
@@ -11,6 +11,8 @@ This file contains the list of changes made to pyjoulescope.
 *   Added optional application-specific metadata to datafile collection start.
 *   Added support for voltage range to datafile: save/load/process correctly.
 *   Improved missing sample (NaN) handling robustness.
+*   Modified Driver.statistics_callback to match statistics_get format.
+*   Compute total charge in addition to energy.
 
 
 ## 0.4.5